Updates after the first workshop delivery
We clean up the pip installs for windows which did not like the
single quotes or the backslashes.

We also add checks for the proper python versions. Students with
Python 3.13 (which is the current release) failed to pip install
torch.

We also remove the notebook cells for installing ollama in colab.
Even those students who planned to use Replicate, still ran the cells
to install ollama in colab which was a poor experience. So we basically
remove the ollama-in-colab information from the workshop.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>